Add metodo selecionarAvaliador() no controlador Artigos configurando routes e views. Alteração do atributo usuario_avaliar para Long.

+++ trunk/congic-ifrn-web/congic-ifrn/conf/evolutions/default/1.sql
@@ -35,7 +35,7 @@ create table artigo (
   objetivos                 longtext,
   metodologia               longtext,
   referencias               longtext,
-  usuario_avaliar           integer,
+  usuario_avaliar           bigint,
   arquivo                   BLOB NOT NULL,
   constraint pk_artigo primary key (id))
 ;
